Move current function into a utility module so it can be re-used after improvement.
